refactor : make code structure to be more flexible for author deconstructor service
