+++ runarion-python/src/models/request.py
@@ -1,6 +1,6 @@
-# model/request.py
+# models/request.py
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 from typing import Optional, Dict
 
 class CallerInfo(BaseModel):
@@ -9,24 +9,7 @@ class CallerInfo(BaseModel):
     project_id: str
     api_keys: Dict[str, Optional[str]]  # e.g., {"openai": "", "gemini": "", "deepseek": "sk-..."}
 
-class PromptConfig(BaseModel):
-    author_profile: Optional[str] = ""
-    context: Optional[str] = ""
-    genre: Optional[str] = ""
-    tone: Optional[str] = ""
-    pov: Optional[str] = ""  # point of view
-
-class GenerationConfig(BaseModel):
-    temperature: float = Field(0.7, ge=0.0, le=1.0)
-    max_output_tokens: int = Field(200, ge=1)
-    top_p: float = Field(1.0, ge=0.0, le=1.0)
-    top_k: float = Field(0.0, ge=0.0)
-    repetition_penalty: float = Field(0.0, ge=0.0)
-
-class GenerationRequest(BaseModel):
-    prompt: Optional[str] = None
+class BaseGenerationRequest(BaseModel):
     provider: str = "openai"
-    model: str
+    model: Optional[str] = None
     caller: CallerInfo
-    prompt_config: PromptConfig
-    generation_config: GenerationConfig
