+++ runarion-python/src/models/response.py
@@ -1,4 +1,6 @@
-from pydantic import BaseModel, Field
+# models/response.py
+
+from pydantic import BaseModel
 from typing import Literal, Optional
 
 class UsageMetadata(BaseModel):
@@ -8,15 +10,13 @@ class UsageMetadata(BaseModel):
     total_tokens: int
     processing_time_ms: int
 
-
 class QuotaMetadata(BaseModel):
     user_id: str
     workspace_id: str
     project_id: str
     generation_count: int
 
-
-class GenerationResponse(BaseModel):
+class BaseGenerationResponse(BaseModel):
     success: bool = True
     text: str
     provider: str
