+++ runarion-python/src/models/story_generation/response.py
@@ -1,0 +1,6 @@
+# models/story_generation/response.py
+
+from models.response import BaseGenerationResponse
+
+class StoryGenerationResponse(BaseGenerationResponse):
+    pass
