+++ runarion-python/src/providers/base_provider.py
@@ -1,36 +1,23 @@
+# providers/base_provider.py
+
 import os
 from abc import ABC, abstractmethod
 from flask import current_app
 from typing import Literal
-from models.request import GenerationRequest
-from models.response import GenerationResponse, UsageMetadata, QuotaMetadata
-from utils.instruction_builder import InstructionBuilder
-from services.quota_manager import QuotaManager
+from models.request import BaseGenerationRequest
+from models.response import BaseGenerationResponse
 
 class BaseProvider(ABC):
-    def __init__(self, request: GenerationRequest):
+    def __init__(self, request: BaseGenerationRequest):
         self.request = request
-        self.instruction = self._build_instruction()
-
-        # Determine which API key is actually used: own or default
-        self.api_key, self.key_used = self._api_key_resolver()
+        self.api_key, self.key_used = self._resolve_api_key()
         self.model = self._resolve_model()
-        self.quota_manager = self._get_quota_manager()
-        
-    def _get_quota_manager(self) -> QuotaManager:
-        return QuotaManager()
 
-    def _check_quota(self):
-        self.remaining_quota = self.quota_manager.check_quota(self.caller)
+    @abstractmethod
+    def generate(self) -> BaseGenerationResponse:
+        pass
 
-    def _update_quota(self, quota_generation_count: int):
-        self.quota_manager.update_quota(
-            caller=self.caller,
-            expected_quota=self.remaining_quota,
-            quota_generation_count=quota_generation_count
-        )
-
-    def _api_key_resolver(self) -> tuple[str, Literal["own", "default"]]:
+    def _resolve_api_key(self) -> tuple[str, Literal["own", "default"]]:
         key = self.request.caller.api_keys.get(self.request.provider)
         if key and key.strip():
             return key.strip(), "own"
@@ -67,7 +54,7 @@ class BaseProvider(ABC):
         env_key = env_map.get(self.request.provider)
         if not env_key:
             raise ValueError(f"No fallback model mapping found for provider: {self.request.provider}")
-        
+
         default_model = os.getenv(env_key)
         if not default_model:
             raise ValueError(
@@ -76,150 +63,3 @@ class BaseProvider(ABC):
 
         current_app.logger.warning("No model name provided in request, using default model.")
         return default_model
-    
-    def _build_instruction(self) -> str:
-        """
-        Build the instruction string based on the prompt configuration.
-        """
-        builder = InstructionBuilder(self.request.prompt_config)
-        instruction = builder.build() if self.request.prompt.strip() else builder.build_from_scratch()
-        return instruction.strip()
-
-    @abstractmethod
-    def generate(self) -> GenerationResponse:
-        """
-        Generate text using the specified provider.
-
-        This method must return a GenerationResponse instance with all required fields filled.
-        """
-        pass
-    
-    def _build_error_response(
-        self,
-        request_id: str,
-        provider_request_id: str = "",
-        error_message: str = "An error occurred during generation.",
-    ) -> GenerationResponse:
-        """Helper to build an error response object."""
-
-        metadata = UsageMetadata(
-            finish_reason="error",
-            input_tokens=0,
-            output_tokens=0,
-            total_tokens=0,
-            processing_time_ms=0,
-        )
-
-        quota = QuotaMetadata(
-            user_id=self.request.caller.user_id,
-            workspace_id=self.request.caller.workspace_id,
-            project_id=self.request.caller.project_id, 
-            generation_count=0,
-        )
-
-        response = GenerationResponse(
-            success=False,
-            text="",
-            provider=self.request.provider,
-            model_used=self.model,
-            key_used=self.key_used,
-            request_id=request_id,
-            provider_request_id=provider_request_id,
-            metadata=metadata,
-            quota=quota,
-            error_message=error_message
-        )
-        return response
-
-    def _build_response(
-        self,
-        generated_text: str,
-        finish_reason: str,
-        input_tokens: int,
-        output_tokens: int,
-        total_tokens: int,
-        processing_time_ms: int,
-        request_id: str,
-        quota_generation_count: int,
-        provider_request_id: str = "",
-    ) -> GenerationResponse:
-        """Helper to build the standardized response object."""
-
-        metadata = UsageMetadata(
-            finish_reason=finish_reason,
-            input_tokens=input_tokens,
-            output_tokens=output_tokens,
-            total_tokens=total_tokens,
-            processing_time_ms=processing_time_ms,
-        )
-
-        quota = QuotaMetadata(
-            user_id=self.request.caller.user_id,
-            workspace_id=self.request.caller.workspace_id,
-            project_id=self.request.caller.project_id,
-            generation_count=quota_generation_count,
-        )
-
-        response = GenerationResponse(
-            success=True,
-            text=generated_text,
-            provider=self.request.provider,
-            model_used=self.model,
-            key_used=self.key_used,
-            request_id=request_id,
-            provider_request_id=provider_request_id,
-            metadata=metadata,
-            quota=quota,
-        )
-        return response
-    
-    def _log_generation_to_db(self, response: GenerationResponse):
-        try:
-            with self.quota_manager.connection_pool.getconn() as conn:
-                with conn.cursor() as cursor:
-                    cursor.execute("""
-                        INSERT INTO generation_logs (
-                            request_id,
-                            user_id,
-                            workspace_id,
-                            project_id,
-                            provider,
-                            model_used,
-                            key_used,
-                            prompt,
-                            instruction,
-                            generated_text,
-                            success,
-                            finish_reason,
-                            input_tokens,
-                            output_tokens,
-                            total_tokens,
-                            processing_time_ms,
-                            error_message,
-                            created_at
-                        ) VALUES (
-                            %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, NOW()
-                        )
-                    """, (
-                        response.request_id,
-                        int(response.quota.user_id),
-                        response.quota.workspace_id,
-                        response.quota.project_id,
-                        response.provider,
-                        response.model_used,
-                        response.key_used,
-                        self.request.prompt or "",
-                        self.instruction or "",
-                        response.text or "",
-                        response.success,
-                        response.metadata.finish_reason,
-                        response.metadata.input_tokens,
-                        response.metadata.output_tokens,
-                        response.metadata.total_tokens,
-                        response.metadata.processing_time_ms,
-                        response.error_message
-                    ))
-                    conn.commit()
-        except Exception as e:
-            current_app.logger.error(f"Failed to log generation to DB: {e}")
-
