+++ runarion-python/src/providers/story_generation/gemini_provider.py
@@ -4,14 +4,13 @@ import time
 import uuid
 from flask import current_app
 from google import genai
-from models.request import GenerationRequest
-from models.response import GenerationResponse
-from providers.base_provider import BaseProvider
-from services.quota_manager import QuotaManager
-from werkzeug.exceptions import Forbidden
+from models.story_generation.request import StoryGenerationRequest
+from models.story_generation.response import StoryGenerationResponse
+from providers.story_generation.base_provider import StoryGenerationBaseProvider
+from google.genai.types import GenerateContentConfig
 
-class GeminiProvider(BaseProvider):
-    def __init__(self, request: GenerationRequest):
+class StoryGenerationGeminiProvider(StoryGenerationBaseProvider):
+    def __init__(self, request: StoryGenerationRequest):
         super().__init__(request)
 
         try:
@@ -19,24 +18,17 @@ class GeminiProvider(BaseProvider):
         except Exception as e:
             current_app.logger.error(f"Failed to initialize Gemini client: {e}")
             raise ValueError(f"Failed to initialize Gemini client: {str(e)}")
-        
-        self.caller = self.request.caller
-        self.quota_manager = self._get_quota_manager()
-        self.remaining_quota = None
 
-    def generate(self) -> GenerationResponse:
-        # Prepare parameters
+    def generate(self) -> StoryGenerationResponse:
         model_to_use = self.model
         prompt = self.request.prompt or "<start writing from scratch>"
         instruction = self.instruction
-
         gen_cfg = self.request.generation_config
-        
-        # Gemini param mapping - adjust as per the SDK you use
+
         gemini_kwargs = {
             "model": model_to_use,
             "contents": prompt,
-            "config": genai.types.GenerateContentConfig(
+            "config": GenerateContentConfig(
                 system_instruction=instruction,
                 temperature=gen_cfg.temperature,
                 max_output_tokens=gen_cfg.max_output_tokens,
@@ -49,7 +41,7 @@ class GeminiProvider(BaseProvider):
         quota_generation_count = 1
         request_id = str(uuid.uuid4())
         provider_request_id = None
-        
+
         try:
             self._check_quota()
         except Exception as e:
@@ -59,28 +51,26 @@ class GeminiProvider(BaseProvider):
                 provider_request_id="" if not provider_request_id else provider_request_id,
                 error_message=f"Gemini Quota error: {str(e)}",
             )
-            
             self._log_generation_to_db(response)
             return response
 
-
         try:
-            # Call the Gemini generate_content endpoint
             raw_response = self.client.models.generate_content(**gemini_kwargs)
             provider_request_id = getattr(raw_response, 'response_id', "")
-            
-            # Check for block reason
+
             if raw_response.prompt_feedback and raw_response.prompt_feedback.block_reason:
                 reason_name = getattr(raw_response.prompt_feedback.block_reason, 'name', str(raw_response.prompt_feedback.block_reason))
                 error_message = f"Content generation blocked by Gemini. Reason: {reason_name}"
                 current_app.logger.warning(error_message)
                 return self._build_error_response(
                     request_id=request_id,
-                    provider_request_id="" if not provider_request_id else provider_request_id,
+                    provider_request_id=provider_request_id,
                     error_message=error_message
                 )
-                
-            # Extract text from first candidate
+
+            generated_text = ""
+            finish_reason = ""
+
             if raw_response.candidates:
                 candidate = raw_response.candidates[0]
                 if candidate.content and candidate.content.parts:
@@ -96,7 +86,7 @@ class GeminiProvider(BaseProvider):
             total_tokens = getattr(usage, "total_token_count", 0)
 
             processing_time_ms = int((time.time() - start_time) * 1000)
-            
+
             self._update_quota(quota_generation_count)
 
             response = self._build_response(
@@ -110,7 +100,7 @@ class GeminiProvider(BaseProvider):
                 provider_request_id=provider_request_id,
                 quota_generation_count=quota_generation_count,
             )
-            
+
             self._log_generation_to_db(response)
             return response
 
@@ -118,9 +108,8 @@ class GeminiProvider(BaseProvider):
             current_app.logger.error(f"Gemini API error with model {model_to_use}: {e}")
             response = self._build_error_response(
                 request_id=request_id,
-                provider_request_id="" if not provider_request_id else provider_request_id,
+                provider_request_id=provider_request_id or "",
                 error_message=f"Gemini API error: {str(e)}",
             )
-            
             self._log_generation_to_db(response)
             return response
